# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Kredete Backend Role Test APIs",
    "description": "Kredete Backend Role Test APIs",
    "version": "0.1.0"
  },
  "host": "127.0.0.1:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
    "/auth/login": {
      "post": {
        "summary": "Authenticate user",
        "description": "",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "payload",
            "in": "body",
            "description": "payload",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string",
                  "default": "you@cool.com"
                },
                "password": {
                  "type": "string",
                  "default": "2azw2x1212@"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The schema below is the value of the data field returned in the response",
            "schema": {
              "type": "object",
              "properties": {
                "accessKey": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Invalid email or password provided"
          }
        }
      }
    },
    "/customers/loans-summary": {
      "get": {
        "summary": "Get customers loan summary",
        "description": "",
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "The schema below is the value of the data field returned in the response",
            "schema": {
              "type": "object",
              "properties": {
                "items": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "customerName": {
                        "type": "string"
                      },
                      "customerId": {
                        "type": "string"
                      },
                      "totalLoanAmount": {
                        "type": "number"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}